pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "formbot"
version = "0.1.0"
description = "Sistema de diligenciamiento inteligente de formularios empresariales."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openpyxl>=3.1.2",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
test = [
  "pytest>=9.0.0",
  "pytest-mock>=3.15.0",
  "httpx>=0.28.0",
]
web = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "python-multipart>=0.0.20",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
